+++ internal/domain/services/plancandidate/change_places_order.go
@@ -2,8 +2,8 @@ package plancandidate
 
 import (
 	"context"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/model"
 
-	"poroto.app/poroto/planner/graphql/model"
 	"poroto.app/poroto/planner/internal/domain/models"
 )
