+++ internal/infrastructure/graphql/factory/price_range.go
@@ -1,8 +1,8 @@
 package factory
 
 import (
-	graphql "poroto.app/poroto/planner/graphql/model"
 	"poroto.app/poroto/planner/internal/domain/models"
+	graphql "poroto.app/poroto/planner/internal/infrastructure/graphql/model"
 )
 
 func PriceRangeFromDomainModel(priceRange *models.PriceRange) *graphql.PriceRange {
