+++ internal/infrastructure/graphql/resolver/schema.resolvers.go
@@ -7,7 +7,7 @@ package resolver
 import (
 	"context"
 
-	"poroto.app/poroto/planner/graphql/generated"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/generated"
 )
 
 // Ping is the resolver for the ping field.
