+++ internal/infrastructure/graphql/resolver/user_query.resolvers.go
@@ -9,9 +9,9 @@ import (
 	"fmt"
 	"log"
 
-	"poroto.app/poroto/planner/graphql/factory"
-	"poroto.app/poroto/planner/graphql/model"
 	"poroto.app/poroto/planner/internal/domain/services/user"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/factory"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/model"
 )
 
 // FirebaseUser is the resolver for the firebaseUser field.
