+++ internal/infrastructure/graphql/schema/image_type.graphqls
