+++ internal/infrastructure/graphql/schema/place_type.graphqls
