+++ internal/infrastructure/graphql/schema/plan_candidate_mutation.graphqls
