+++ internal/infrastructure/graphql/schema/plan_candidate_query.graphqls
