+++ internal/infrastructure/graphql/schema/plan_candidate_type.graphqls
