+++ internal/infrastructure/graphql/schema/plan_query.graphqls
