+++ internal/infrastructure/graphql/schema/plan_type.graphqls
