+++ internal/infrastructure/graphql/schema/user_query.graphqls
