+++ internal/infrastructure/graphql/schema/user_type.graphqls
