+++ internal/interface/rest/graphql.go
@@ -4,8 +4,8 @@ import (
 	"github.com/99designs/gqlgen/graphql/handler"
 	"github.com/99designs/gqlgen/graphql/playground"
 	"github.com/gin-gonic/gin"
-	"poroto.app/poroto/planner/graphql/generated"
-	"poroto.app/poroto/planner/graphql/resolver"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/generated"
+	"poroto.app/poroto/planner/internal/infrastructure/graphql/resolver"
 )
 
 func GraphQlPlayGround(c *gin.Context) {
